requirements.txt
kafka-python==1.3.1
